rename erc1271-verifier -> scw_verifiers
